// lib/src/drawing/sprite-modifier-group.d.ts
import { SpriteModifier } from './sprite-modifier';
import { BSpriteGroup } from '../b-export/b-building';
export declare class SpriteModifierGroup {
    groupName: string;
    spriteModifiers: SpriteModifier[];
    constructor();
    importFrom(original: BSpriteGroup): void;
}
//# sourceMappingURL=sprite-modifier-group.d.ts.map