// lib/src/enums/build-location-rule.d.ts
export declare enum BuildLocationRule {
    Anywhere = 0,
    OnFloor = 1,
    OnFloorOverSpace = 2,
    OnCeiling = 3,
    OnWall = 4,
    InCorner = 5,
    Tile = 6,
    NotInTiles = 7,
    Conduit = 8,
    LogicBridge = 9,
    WireBridge = 10,
    HighWattBridgeTile = 11,
    BuildingAttachPoint = 12,
    OnFloorOrBuildingAttachPoint = 13,
    OnFoundationRotatable = 14
}
//# sourceMappingURL=build-location-rule.d.ts.map